combining samples

+++ train_EEG.py
@@ -611,7 +611,7 @@ def run_train_EC_Multi(learn_rate, wd, batch_sz, epochs, outfile):
             print(f"label: {label}")
             print(f"shape: {label.shape}\n")
             label_reshaped = np.reshape(label, (batch_sz,1,5))
-            label_reshaped = label_reshaped.type(torch.LongTensor)
+            label_reshaped = label_reshaped.type(torch.float32)
             print(f"label: {label_reshaped}")
             print(f"shape: {label_reshaped.shape}\n")
